refactor(Playlist): playlist로 url 매핑 컨벤션 통일
